requirements.txt
python-chess==1.999
stockfish==3.28.0
